requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
pinecone-client
pydantic
